Added Validation.
Registraion refactored. Authorization Refactored.
