+++ app/config/routing.yml
@@ -1,3 +1,7 @@
+authentication_registration:
+    resource: "@AuthenticationRegistrationBundle/Resources/config/routing.yml"
+    prefix:   /auth
+
 app:
     resource: "@AppBundle/Controller/"
