+++ app/config/security.yml
@@ -3,7 +3,7 @@
 security:
 
     encoders:
-        AppBundle\Entity\User:
+        Authentication\RegistrationBundle\Entity\User:
             algorithm: plaintext
 
     # http://symfony.com/doc/current/book/security.html#where-do-users-come-from-user-providers
@@ -13,7 +13,7 @@ security:
 
         our_db_provider:
                     entity:
-                        class: AppBundle:User
+                        class: AuthenticationRegistrationBundle:User
                         property: username
                         # if you're using multiple entity managers
                         # manager_name: customer
