+++ src/Authentication/RegistrationBundle/AuthenticationRegistrationBundle.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Authentication\RegistrationBundle;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class AuthenticationRegistrationBundle extends Bundle
+{
+}
