+++ src/Authentication/RegistrationBundle/Controller/DefaultController.php
@@ -1,18 +1,20 @@
 <?php
 
-namespace AppBundle\Controller;
+namespace Authentication\RegistrationBundle\Controller;
 
-use AppBundle\Entity\User;
+use Authentication\RegistrationBundle\Entity\User;
+use Exception;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-use Symfony\Component\Config\Definition\Exception\Exception;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
-class AuthenticationController extends Controller
+class DefaultController extends Controller
 {
     public function registrationAction(Request $request)
     {
         try{
+            $validator = $this->get('validator');
+
             $username = $request->get('username');
             if($this->getDoctrine()->getRepository('AppBundle:User')->findBy(['username' => $username]) == []){
                 $user = new User();
@@ -20,6 +22,12 @@ class AuthenticationController extends Controller
                     ->setPassword($request->get('password'))
                     ->setName($request->get('name'))
                     ->setRoles(['ROLE_CUSTOMER']);
+
+                $validationErrors = $validator->validate($user);
+
+                if(count($validationErrors) > 0) {
+                    return new Response((string)$validationErrors, Response::HTTP_BAD_REQUEST);
+                }
 
                 $em = $this->getDoctrine()->getEntityManager();
                 $em->persist($user);
