+++ src/Authentication/RegistrationBundle/Resources/config/doctrine/User.orm.xml
@@ -4,7 +4,7 @@
                   xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
         http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
 
-    <entity name="AppBundle\Entity\User" table="users">
+    <entity name="Authentication\RegistrationBundle\Entity\User" table="users">
         <id name="id" type="integer">
             <generator strategy="AUTO" />
         </id>
