+++ src/Authentication/RegistrationBundle/Resources/config/routing.yml
@@ -1,0 +1,3 @@
+authentication_registration_homepage:
+    path:     /regitration
+    defaults: { _controller: AuthenticationRegistrationBundle:Default:registration }
