+++ src/Authentication/RegistrationBundle/Resources/config/services.yml
@@ -1,0 +1,4 @@
+services:
+#    authentication_registration.example:
+#        class: Authentication\RegistrationBundle\Example
+#        arguments: ["@service_id", "plain_value", %parameter%]
