+++ src/Authentication/RegistrationBundle/Resources/views/Default/index.html.twig
@@ -1,0 +1,1 @@
+Hello World!
